refactor(theme-classic): enforce collocated JSX and module CSS
