+++ packages/docusaurus-theme-classic/src/theme/DocPage/Layout/Main/styles.module.css
